requirements.txt
# pip install -r requirements.txt
opencv-python
selenium
flask
selenium
transformers
numpy
bs4
requests
Pillow
wcag-contrast-ratio
colorthief
torch